requirements.txt
arrow==1.3.0
bqplot==0.12.44

cligj==0.7.2
colour==0.1.5

earthengine-api==1.5.0
eerepr==0.1.0

fqdn==1.5.1

future==1.0.0

geemap==0.35.1
geocoder==1.38.1
geographiclib==2.0

geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0

httplib2==0.22.0

ipyevents==2.0.2
ipyfilechooser==0.6.0

ipyleaflet==0.19.2

ipytree==0.2.2

isoduration==20.11.0

jsonpointer==3.0.0

jupyter-leaflet==0.19.2

matplotlib==3.9.2


narwhals==1.24.1


proto-plus==1.26.0
protobuf==4.25.3

pyasn1==0.6.1
pyasn1_modules==0.4.1

pyperclip==1.9.0

pyshp==2.3.1

python-box==7.3.2

ratelim==0.1.6

rsa==4.9


soundfile==0.13.1

streamlit-geolocation==0.0.10
streamlit-js-eval==0.1.7
streamlit_folium==0.24.0

traittypes==0.2.1

types-python-dateutil==2.9.0.20241206

uri-template==1.3.0
uritemplate==4.1.1

webcolors==24.11.1
webencodings==0.5.1
pytorch
transformers
geopandas
